+++ lua/mini/bufremove.lua
@@ -84,9 +84,6 @@ end
 --- Default values:
 ---@eval return MiniDoc.afterlines_to_code(MiniDoc.current.eval_section)
 MiniBufremove.config = {
-  -- Whether to set Vim's settings for buffers (allow hidden buffers)
-  set_vim_settings = true,
-
   -- Whether to disable showing non-error feedback
   silent = false,
 }
@@ -189,19 +186,12 @@ H.setup_config = function(config)
   H.check_type('config', config, 'table', true)
   config = vim.tbl_deep_extend('force', vim.deepcopy(H.default_config), config or {})
 
-  H.check_type('set_vim_settings', config.set_vim_settings, 'boolean')
   H.check_type('silent', config.silent, 'boolean')
 
   return config
 end
 
-H.apply_config = function(config)
-  MiniBufremove.config = config
-
-  if config.set_vim_settings then
-    vim.o.hidden = true -- Allow hidden buffers
-  end
-end
+H.apply_config = function(config) MiniBufremove.config = config end
 
 H.is_disabled = function() return vim.g.minibufremove_disable == true or vim.b.minibufremove_disable == true end